Arreglo de doble controlador
